--- pages/callbacks/__init__.py
@@ -1,1 +1,0 @@
-# empty file to make this a module
